hardcode bibutils path
commit [3f13e3ba39ff210a3b20d025c4c6c80ec29e116f] updated the
bibutils source files, but it appears that the production server
is still running the old executable.

this commit hardcodes the location of the bibutils executables,
so that being able to find them does not depend on the user's PATH.

this commit also compiles bibutils automatically if the executables
are not found.