Automatic projection of curly braces in BibTeX entries (for protecting fixed-case words) into XML as <fixed-case> tags.
- `{\cmd ...}` does _not_ protect case, and even braces within don't protect case (this is how BibTeX apparently works)
- `{$...$}` is not tagged, because I figure the conversion back to BibTeX can add the curly braces easily
- In `\cmd{...}`, the argument is not tagged, because I didn't see any cases where it was clearly intended to protect case.

The script asks for manual intervention when the XML has been changed from the BibTeX; I'll make those changes later.

To protect a word like `English`, one is supposed to write `{E}nglish`, but a lot of people (myself included) write `{English}`. The reason is that if a bib style wanted to make the title all-caps, the `nglish` should not be protected. It would be very easy to modify the script to interpret `{English}` as `{E}nglish`, but there are some edge cases, like a system named `{fokas}` where the lowercasing apparently _should_ be protected. So the script currently follows the curly braces strictly, but let me know if you want the other behavior.