new Makefile, copy static files from other directory
Files previously residing in hugo/ are now copied
as part of the build process, the website is built into build/.
This way, rm -rf build cleans everything.

The website to be published is generated into the directory
build/anthology by default, reflecting the directory name used on the
upstream server.  This directory can be configured using
make ANTHOLOGYDIR=someotherdir.

The Makefile consists of a set of phony targets, each building a part
of the site.

Relevant targets:
 - site: creates the site
 - all: cleans output, checks validity of XML and creates the site
 - serve: starts a web server on localhost:8000 to browse the site
 - upload: rsyncs the site to the aclweb webserver