Makefile: first parallelizations, remove progress bars
Changed the targets to be empty targets instead of phony ones (phony
ones are still there for a nicer interface on the command line), model
the correct dependencies between them so that make -j4 works.

Removed the progress bars from bibtex etc. generation as it is not
printed correctly anyways when doing a parallel build.