fixed main volume book title
I'm not sure how this was set correctly. I chose this name for consistency with [past volumes](https://www.aclweb.org/anthology/venues/ranlp/), but have suggestions and have reached out to the volume editor to change it further. However, I think we should adopt this variant right away since it would be good to have the correction in and this is clearly a step in the right direction.

I have also created a [rudimentary ingestion checklist](https://github.com/acl-org/acl-anthology/wiki/Ingestion-Checklist) to help me avoid overlooking this kind of mistake in the future.