Mirroring infrastructure
This commit adds
 - a script to download all ACL files not in the git repo
 - a configurable websites directory, including the ability
   to host in any subdir (/anthology/, ... also top-level)
 - reworked anthology-files directory, symlinked into the
   anthology web directory (and automatically adapted in the
   .htaccess file)
 - renamed constants in anthology/data.py, including facility
   to set them via environment variables
 - additional Makefile documentation
 - checks in the Makefile and fewer dependencies on phony tasks

create_mirror.py reads a list of anthology XML files, checks the
checksums of already downloaded files, downloads new ones and checks
the checksum to only put correct files into the download dir.  It can
be parallelized by calling it several times with different sets of XML
files.

The website is now built under a "website" directory; otherwise
top-level builds would not be separate from other generated data.  The
Makefile creates a symlink inside the anthology directory to the path
where the anthology-files will be on the server.  Apache needs to
follow symlinks for this to work.

The environment variable ANTHOLOGY_PREFIX defines the host and
directory under which the site is supposed to be hosted.

The constants in anthology/data.py have been renamed (long-standing
TODO) and the canonical URL template has been separated from the
host and prefix used for hosting a copy.  It is therefore now possible
to host a mirror of only the HTML or HTML plus files.