Don't need ID if name is unambiguous
If a slugified name appears in name_variants.yaml with one ID,
and that ID is only used some of the time, we now assume that
it is the same person and do not give an error.  If there are
two IDs for the same slugified name, you must include the ID
in every paper.

All IDs used in the xml files must appear in name_variants.yaml.
If we start getting lots of ORCIDs in new data, we will need some
tool for adding them to name_variants.yaml.