Explicit representation of events (#1896)
This fairly large change moves from an implicit to an explicit representation of events and the volume-venue association. Among the major changes:

* Every volume now explicitly denotes its venue using a `<venue>` tag in its meta block. This eliminates complicated logic dealing with old-style IDs. Multiple tags are allowed to represent joint events.

* An event is created for every collection, which remains defined as a (venue, year) pairing. These events comprise all volumes having a venue tag associating them with that venue.

* An `<event>` block within each collection is used to list all colocated volumes.

Together, these replace the `joint.yaml` file which was previously used to encode colocated volumes and joint events, implementing many of the ideas in #1164. This also paves the way for adding event-related items, such as keynotes, the handbook, and so on, to each event page (#298).

Technically:
* The `Anthology` class now builds an explicit representation of all events.
* Every `Volume` now knows its venues. Complicated logic dealing with old-style IDs been eliminated, and can probably be further simplified. 
* Venues are now sorted and keyed by slug, rather than acronym.

Helper scripts `normalize_joint_yaml.py`, `split_joint.py`, and `split_venues.py` were used to help with the conversion.

In the process, many minor corrections were made to papers and events.

Co-authored-by: Marcel Bollmann <marcel@bollmann.me>