Remove AnthologyIndex.verify()
Removing `AnthologyIndex.verify()` because it doesn't actually do what it
advertises. The comment suggests that it throws an error when a name is used
both without and with more than one explicit ID; what it actually checks is if
the paper is used more than once with an explicit ID, even if that ID is always
the same. The ambiguity check should already be performed in
`AnthologyIndex.register()`, so I don't see a reason to keep this function
around at all, because all it does right now is raise false positives.