Auto pr for paper information correction (#4138)
This change contains the following features:

On each paper page, a collapsible is added at the bottom, with the indication "File a correction". When clicked on, it shows a form prefilled current of the paper
Users can modify these information and click submit
After submission, the user is taken to a window that creates pull request to the repository. The effect of the pr is to add file containing these new information that the user entered. The path to the file and the branch that the pr is aiming can be specified.